Fix memory leak in render_overlap
render_overlap allocated memory for its hashmap key and values on the
heap, relying on the cache cleanup to free them.  However, these
pointers are not directly inserted into the cache, but memcpy()'ed in
hashmap_insert, leading to a pretty bad memory leak.
Allocate the key and value on the stack instead to fix this problem.