Support \xbord and \ybord override tags
Improve get_outline_glyph with a second pass which can be used for
stroking with different width in x and y direction.  This is done
by first stroking in y direction, and if the stroke size for the x
direction differs, stroking a copy of the glyph in x direction again.
The x coordinates of the first stroker result are afterwards replaced
by the coordinates of the second stroking.
The cache hash keys for bitmaps and glyphs have been extended to carry
the new properties.