Fix two-pass stroking for \xbord, \ybord
The two-pass stroker blindly assumed that the number of points of
the stroked glyph is the same, no matter the size of the stroker.
Unfortunately, this is not the case every time.  In such cases, the
coordinates will only be replaced up to mininum of points of both
glyphs.  It's incredibly hacky, but seems to work well (and look good).