Support for \iclip (inverse clipping)
Inverse clipping means to display everything outside the clipping
rectangle.  It is implemented by splitting up a bitmap into up
to 4 parts, as needed by the clipping rectangle.  Afterwards, the
bitmap is clipped against the screen boundaries as usual.  Finally,
each bitmap is split up into left and right and colored different,
in case \kf karaoke is in effect.