Scale rotation constants with font scale
Rotations use a "magic" constant of, by default, 20000.  This number
is more or less taken from VSFilter; I'm not even really sure what is
(viewing distance, maybe), but scaling it according to font_scale
solves problems with distortions of rotations in case the rendering
resolution differs from the video resolution.