Fix crash when event format spec is missing
libass didn't properly initialize its idea about the event format
specification (the Format: line in the [Events] section) in case none
was explicitly specified.  This ends in a crash due to access to a null
pointer.  It was changed so that the event format is initialized to
a reasonable default, the way it is also done for embedded subtitles.