Use better method for stroking with \xbord, \ybord
Instead of stroking in two passes using the regular stroker use
the outline emboldener in case different x and y stroke widths are
requested.

The regular outline stroker draws circular arcs, adding points to
the outline.  The number of points can vary according to the width
of the stroker.  This makes it impossible to reliably stroke in
two passes with it.
The outline emboldener does not produce outlines looking as nice
as the stroker outlines, but it never adds points to the outline,
making it suitable for two-pass emboldening with different x and y
width.