Redo border calculations for \be and \blur
The last commit related to this overdid it a bit -- in some cases,
with \be5, for example, edges were visible on the outlines.  Instead of
using a linear function to map \be applications to border enlargement
value, use a nonlinear function, a simple square root. This seems to
work a lot better, while keeping the enlargement small.
In addition, make it possible to use \be and \blur at the same time. \be
is applied first.