Implement vector clips
Make it possible to use drawings for clipping with \clip and \iclip.
parse_tag was extended to parse drawings in \clip or \iclip tags in case
parsing them as rectangular clips fails.  These clip drawings are later
rasterized and used for blending bitmaps, just after bitmaps are
assembled into a list in render_text.  Currently, the cache is not
utilized for storing the parsed drawings or blended bitmaps.