Render outlines, not glyphs
libass can render both regular glyphs (extracted from a font) and
drawings, which are inserted into an existing glyph so far. This is a
rather crude hack. However, it is necessary, as the renderer only works
with FT_Glyphs. The obvious solution is to work on the underlying
outlines directly instead.

This is the first step to make this possible: change the basic
rasterizer to use the outline-specific rasterization functions. There
are more advantages:

- It's more efficient. The bitmap can be directly rendered into the
  bordered target bitmap.
- This should make it easier to adapt the renderer to run-based
  rendering later on (the pixels are ORed into the target)

The higher-level rendering code still needs to be adjusted.