Separate glyph font face matching and glyph loading
Split up ass_font_get_glyph into two functions: one function matches
checks the fonts for support of the requested codepoint and returns
a face_index and glyph_index to later load the glyph, with a second
function.

This approach is very useful for shaping, which needs to be done
on runs with the same font face and font size and changes the
glyph_index.