Reset text origin on \fay changes for VSFilter compatibility
Comparing this rendering of this line:

{\fnTahoma\c&H000000&\fs100\an7\fay0.115\pos(240,250)}——————–>{\fay0.0}——>

To this line:

{\fnTahoma\c&H000000&\fs100\an7\fay0.0\pos(240,250)}——————–>{\fay0.0}——>

The second arrow is in the same place for both lines when using
VSFilter, but the origin of the second shear is different when using
libass.

After every change in \fay or a newline, the vertical offset resets
to 0.

Fixes issue #80. Bug report and patch by BwackNinja (BwackNi...@gmail.com)
(this commit corresponds to the patch issue_80_3.patch)