Duplicate event parsing code for speed
ass_process_chunk() actually can show up in the profiler. It does a lot
of work to handle custom "Format: " lines, which are basically an
inoffical libass extension, and is not widely used. (It's also usedto
handle SSA.)

Keep the parsing code process_event_tail() for compatibility, and parse
modern ASS events directly.