Fix overflow in strtod function
The exponent for scientific notation (e.g. 23E10) is stored as an int,
which can easily overflow. This results in an incorrect conversion,
which we may or may not care about as it requires construction of
floating point number strings with invalid exponents.

The strtod algorithm flips the exponent sign from negative to positive
at some point, which can become a problem because of two's complement
representation. The largest negative number ("weird number") does not
have a counterpart with positive sign. So the result of flipping the
sign will remain negative and a check for a maximum allowed exponent
will fail.

Fix this by adding an additional check for the exponent sign, which
will clamp the exponent to the maximum in these situations.

This fixes libass#244. Whether we want to keep using this strtod
implementation needs to be discussed further.